scope setup on scope services
moved ScopeSetup and related classes to file scope.py under services module.

frame_save_parameter.py when initialized get an instance of ScopeSetup and streams it.

lnet instead, has a real instance of ScopeSetup and feeds it into save_parameter when serializing.

ScopeSetup should not be regenerated and the instance should be called from lnet object.

next step is to move channel related stuff to inside ScopeSetup class and
automate it when calling load_parameter